Fix issue with the EventQueryResultToTypeMapper
If a property of the mapped object (marked with
`ChronologicalEventField`) was of type `Nullable<double>` or
`Nullable<DateTime>`, its value was never set correctly.

This commit fixes this issue and allows to use nullable types.

I've also converted two `IEnumerable` to arrays so that they are not
enumerated multiple times.